pyproject.toml
[project]
name = "minew-msr-01"
version = "0.1.0"
description = "Minew MSR-01 gateway to a webhook or DB"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anyio>=4.9.0",
    "bleak>=0.22.3",
    "coloredlogs>=15.0.1",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "python-dotenv>=1.1.1",
    "structlog>=25.4.0",
]
